refactor(kg): split oversized graph modules after review
- Extract dashboard queries, retention SQL, query row types, and entity filtering into focused co-located modules.

- Preserve the existing public exports and KnowledgeGraphEntity method signatures.

- Include retentionDays in maintenance logs and cover both observation and relation windows.
