+++ src/mcp/tools/kg-archivist/entity-filter.ts
@@ -1,0 +1,240 @@
+// Lexical filters used by NLP entity extraction.
+const PRONOUNS = new Set([
+	"i",
+	"me",
+	"my",
+	"myself",
+	"we",
+	"us",
+	"our",
+	"ours",
+	"ourselves",
+	"you",
+	"your",
+	"yours",
+	"yourself",
+	"yourselves",
+	"he",
+	"him",
+	"his",
+	"himself",
+	"she",
+	"her",
+	"hers",
+	"herself",
+	"it",
+	"its",
+	"itself",
+	"they",
+	"them",
+	"their",
+	"theirs",
+	"themselves",
+	"this",
+	"that",
+	"these",
+	"those",
+	"someone",
+	"somebody",
+	"something",
+	"anyone",
+	"anybody",
+	"anything",
+	"everyone",
+	"everybody",
+	"everything",
+	"nobody",
+	"nothing"
+]);
+
+/** Common English stopwords unlikely to be meaningful "concept" entities. */
+const STOPWORDS = new Set([
+	"a",
+	"an",
+	"the",
+	"and",
+	"but",
+	"or",
+	"if",
+	"because",
+	"as",
+	"until",
+	"while",
+	"of",
+	"at",
+	"by",
+	"for",
+	"with",
+	"about",
+	"against",
+	"between",
+	"into",
+	"through",
+	"during",
+	"before",
+	"after",
+	"above",
+	"below",
+	"to",
+	"from",
+	"up",
+	"down",
+	"in",
+	"out",
+	"on",
+	"off",
+	"over",
+	"under",
+	"again",
+	"further",
+	"then",
+	"once",
+	"here",
+	"there",
+	"when",
+	"where",
+	"why",
+	"how",
+	"all",
+	"each",
+	"every",
+	"both",
+	"few",
+	"more",
+	"most",
+	"other",
+	"some",
+	"such",
+	"no",
+	"nor",
+	"not",
+	"only",
+	"own",
+	"same",
+	"so",
+	"than",
+	"too",
+	"very",
+	"just",
+	"also",
+	"any",
+	"thing",
+	"things",
+	"way",
+	"ways",
+	"person",
+	"people",
+	"man",
+	"woman",
+	"child",
+	"time",
+	"year",
+	"day",
+	"number",
+	"world",
+	"life",
+	"hand",
+	"part",
+	"place",
+	"case",
+	"week",
+	"company",
+	"system",
+	"program",
+	"work",
+	"group",
+	"problem",
+	"fact",
+	"example",
+	"member",
+	"car",
+	"city",
+	"state",
+	"country",
+	"area",
+	"water",
+	"air",
+	"money",
+	"data",
+	"information",
+	"software",
+	"code",
+	"file",
+	"server",
+	"database",
+	"application",
+	"user",
+	"users",
+	"project",
+	"task",
+	"memory",
+	"value",
+	"name",
+	"type",
+	"list",
+	"set",
+	"number",
+	"id",
+	"key",
+	"text",
+	"content",
+	"title",
+	"description",
+	"status",
+	"time",
+	"dan",
+	"yang",
+	"di",
+	"ke",
+	"dari",
+	"dengan",
+	"ini",
+	"itu",
+	"ada",
+	"akan",
+	"bisa",
+	"telah",
+	"sudah",
+	"juga",
+	"atau",
+	"karena",
+	"untuk",
+	"pada",
+	"sebagai",
+	"oleh",
+	"saat",
+	"setelah",
+	"antara",
+	"tentang"
+]);
+
+const DETERMINERS = /^(a|an|the)\s+/i;
+const TRAILING_PUNCTUATION = /[.,!?;:()"'[\]]+$/g;
+const ENTITY_NAME_BAD_PATTERN = /^(~|[-→·•])|["'`[\]{}()]|→|=>|->|`|~[\d]/;
+const ENTITY_NAME_ONLY_SYMBOLS = /^[^a-zA-Z0-9]+$/;
+const ENTITY_NAME_HEADING = /^#/;
+const ENTITY_NAME_ORDINAL_LIST = /^\d+[.)]\s/;
+
+export function cleanText(raw: string): string {
+	return raw.replace(TRAILING_PUNCTUATION, "").trim();
+}
+
+export function stripLeadingDeterminer(phrase: string): string {
+	return phrase.replace(DETERMINERS, "").trim();
+}
+
+export function isExcludedNoun(candidate: string): boolean {
+	const lower = candidate.toLowerCase();
+	return lower.length < 2 || PRONOUNS.has(lower) || STOPWORDS.has(lower) || /^\d+$/.test(candidate);
+}
+
+/** Reject code fragments, punctuation-only text, and markdown structure. */
+export function isValidEntityName(name: string): boolean {
+	return (
+		name.length >= 2 &&
+		!ENTITY_NAME_ONLY_SYMBOLS.test(name) &&
+		!ENTITY_NAME_BAD_PATTERN.test(name) &&
+		!ENTITY_NAME_HEADING.test(name) &&
+		!ENTITY_NAME_ORDINAL_LIST.test(name)
+	);
+}
