+++ src/mcp/tools/kg-archivist/extract.ts
@@ -3,6 +3,7 @@ import { SQLiteStore } from "../../storage/sqlite";
 import { logger } from "../../utils/logger";
 import { KG_MAX_COOCCURRENCE_ENTITIES } from "../../utils/constants";
 import { KgObservationDomain, observationText } from "./observation-text";
+import { cleanText, isExcludedNoun, isValidEntityName, stripLeadingDeterminer } from "./entity-filter";
 
 // ---------------------------------------------------------------------------
 // Constants
@@ -10,295 +11,9 @@ import { KgObservationDomain, observationText } from "./observation-text";
 
 const MAX_CONTENT_LENGTH = 5000;
 
-const PRONOUNS = new Set([
-	"i",
-	"me",
-	"my",
-	"myself",
-	"we",
-	"us",
-	"our",
-	"ours",
-	"ourselves",
-	"you",
-	"your",
-	"yours",
-	"yourself",
-	"yourselves",
-	"he",
-	"him",
-	"his",
-	"himself",
-	"she",
-	"her",
-	"hers",
-	"herself",
-	"it",
-	"its",
-	"itself",
-	"they",
-	"them",
-	"their",
-	"theirs",
-	"themselves",
-	"this",
-	"that",
-	"these",
-	"those",
-	"someone",
-	"somebody",
-	"something",
-	"anyone",
-	"anybody",
-	"anything",
-	"everyone",
-	"everybody",
-	"everything",
-	"nobody",
-	"nothing"
-]);
-
-/** Common English stopwords unlikely to be meaningful "concept" entities. */
-const STOPWORDS = new Set([
-	"a",
-	"an",
-	"the",
-	"and",
-	"but",
-	"or",
-	"if",
-	"because",
-	"as",
-	"until",
-	"while",
-	"of",
-	"at",
-	"by",
-	"for",
-	"with",
-	"about",
-	"against",
-	"between",
-	"into",
-	"through",
-	"during",
-	"before",
-	"after",
-	"above",
-	"below",
-	"to",
-	"from",
-	"up",
-	"down",
-	"in",
-	"out",
-	"on",
-	"off",
-	"over",
-	"under",
-	"again",
-	"further",
-	"then",
-	"once",
-	"here",
-	"there",
-	"when",
-	"where",
-	"why",
-	"how",
-	"all",
-	"each",
-	"every",
-	"both",
-	"few",
-	"more",
-	"most",
-	"other",
-	"some",
-	"such",
-	"no",
-	"nor",
-	"not",
-	"only",
-	"own",
-	"same",
-	"so",
-	"than",
-	"too",
-	"very",
-	"just",
-	"also",
-	"any",
-	"thing",
-	"things",
-	"way",
-	"ways",
-	"person",
-	"people",
-	"man",
-	"woman",
-	"child",
-	"time",
-	"year",
-	"day",
-	"number",
-	"world",
-	"life",
-	"hand",
-	"part",
-	"place",
-	"case",
-	"week",
-	"company",
-	"system",
-	"program",
-	"work",
-	"group",
-	"problem",
-	"fact",
-	"example",
-	"member",
-	"car",
-	"city",
-	"state",
-	"country",
-	"area",
-	"water",
-	"air",
-	"money",
-	"data",
-	"information",
-	"software",
-	"code",
-	"file",
-	"server",
-	"database",
-	"application",
-	"user",
-	"users",
-	"project",
-	"task",
-	"memory",
-	"value",
-	"name",
-	"type",
-	"list",
-	"set",
-	"number",
-	"id",
-	"key",
-	"text",
-	"content",
-	"title",
-	"description",
-	"status",
-	"time",
-	"dan",
-	"yang",
-	"di",
-	"ke",
-	"dari",
-	"dengan",
-	"ini",
-	"itu",
-	"ada",
-	"akan",
-	"bisa",
-	"telah",
-	"sudah",
-	"juga",
-	"atau",
-	"karena",
-	"untuk",
-	"pada",
-	"sebagai",
-	"oleh",
-	"saat",
-	"setelah",
-	"antara",
-	"tentang"
-]);
-
-const DETERMINERS = /^(a|an|the)\s+/i;
-
-const TRAILING_PUNCTUATION = /[.,!?;:()"'[\]]+$/g;
-
-// ---------------------------------------------------------------------------
-// Types
-// ---------------------------------------------------------------------------
-
 export interface ExtractedEntity {
 	name: string;
 	type: "person" | "place" | "organization" | "concept";
-}
-
-// ---------------------------------------------------------------------------
-// Helpers
-// ---------------------------------------------------------------------------
-
-/**
- * Clean trailing punctuation from extracted entity text.
- */
-function cleanText(raw: string): string {
-	return raw.replace(TRAILING_PUNCTUATION, "").trim();
-}
-
-/**
- * Remove leading determiners ("a ", "an ", "the ") from a noun phrase.
- */
-function stripLeadingDeterminer(phrase: string): string {
-	return phrase.replace(DETERMINERS, "").trim();
-}
-
-/**
- * Check whether a noun-phrase candidate should be excluded from concept
- * extraction (pronouns, stopwords, too short, etc.).
- */
-function isExcludedNoun(candidate: string): boolean {
-	const lower = candidate.toLowerCase();
-	if (lower.length < 2) return true;
-	if (PRONOUNS.has(lower)) return true;
-	if (STOPWORDS.has(lower)) return true;
-	if (/^\d+$/.test(candidate)) return true;
-	return false;
-}
-
-const ENTITY_NAME_BAD_PATTERN = /^(~|[-→·•])|["'`[\]{}()]|→|=>|->|`|~[\d]/;
-const ENTITY_NAME_ONLY_SYMBOLS = /^[^a-zA-Z0-9]+$/;
-
-/**
- * Markdown structure that `compromise` happily reports as a noun phrase but
- * which carries no domain meaning (audit F8). These are the highest-degree
- * nodes in a real graph precisely BECAUSE they are boilerplate: the same
- * task/memory template heading appears in hundreds of documents, so it
- * co-occurs with everything and links nothing.
- *
- * Measured on a 45,160-entity corpus: `'### 3. Acceptance & Verification'`
- * was the single highest-degree node at 13,787 edges, and 718 heading-shaped
- * entities (1.6% of all entities) anchored 113,032 edges (8.8% of the graph).
- *
- *   - `HEADING`      — a leading `#` (ATX heading, or a `#RRGGBB` color hex)
- *   - `ORDINAL_LIST` — a leading `1. ` / `2) ` enumerator
- *
- * Both are structural markers, never entity names. Prose that legitimately
- * starts with a digit (`"3D rendering"`, `"2FA"`) has no `.`/`)` separator
- * and is unaffected.
- */
-const ENTITY_NAME_HEADING = /^#/;
-const ENTITY_NAME_ORDINAL_LIST = /^\d+[.)]\s/;
-
-/**
- * Reject entity names that are clearly garbage: code fragments, size
- * references, quote/bracket pollution, pure-symbol strings, or markdown
- * structure (headings / ordinal list markers — audit F8).
- */
-function isValidEntityName(name: string): boolean {
-	if (name.length < 2) return false;
-	if (ENTITY_NAME_ONLY_SYMBOLS.test(name)) return false;
-	if (ENTITY_NAME_BAD_PATTERN.test(name)) return false;
-	if (ENTITY_NAME_HEADING.test(name)) return false;
-	if (ENTITY_NAME_ORDINAL_LIST.test(name)) return false;
-	return true;
 }
 
 // ---------------------------------------------------------------------------
